Add prediction for Tech Disks, cleanup
