+++ Content.Server/Research/Systems/ResearchSystem.Technology.cs
@@ -132,25 +132,6 @@ public sealed partial class ResearchSystem
     }
 
     /// <summary>
-    /// Adds a lathe recipe to the specified technology database
-    /// without checking if it can be unlocked.
-    /// </summary>
-    public void AddLatheRecipe(EntityUid uid, string recipe, TechnologyDatabaseComponent? component = null)
-    {
-        if (!Resolve(uid, ref component))
-            return;
-
-        if (component.UnlockedRecipes.Contains(recipe))
-            return;
-
-        component.UnlockedRecipes.Add(recipe);
-        Dirty(uid, component);
-
-        var ev = new TechnologyDatabaseModifiedEvent();
-        RaiseLocalEvent(uid, ref ev);
-    }
-
-    /// <summary>
     ///     Returns whether a technology can be unlocked on this database,
     ///     taking parent technologies into account.
     /// </summary>
