+++ Content.Shared/Research/Systems/SharedResearchSystem.cs
@@ -280,4 +280,23 @@ public abstract class SharedResearchSystem : EntitySystem
         comp.UnlockedTechnologies.Clear();
         Dirty(uid, comp);
     }
+
+    /// <summary>
+    /// Adds a lathe recipe to the specified technology database
+    /// without checking if it can be unlocked.
+    /// </summary>
+    public void AddLatheRecipe(EntityUid uid, string recipe, TechnologyDatabaseComponent? component = null)
+    {
+        if (!Resolve(uid, ref component))
+            return;
+
+        if (component.UnlockedRecipes.Contains(recipe))
+            return;
+
+        component.UnlockedRecipes.Add(recipe);
+        Dirty(uid, component);
+
+        var ev = new TechnologyDatabaseModifiedEvent();
+        RaiseLocalEvent(uid, ref ev);
+    }
 }
